data moved from App to Router, folder routers moved, index.ts for constants deleted
