+++ src/App.tsx
@@ -1,17 +1,14 @@
 import React from 'react';
-import { useRoutes } from 'react-router-dom';
-import { Routes } from './components/Routes';
 import { Router } from './components/Router';
 import { Header } from './components/Header';
 import { ThemeProvider } from '@mui/material';
 import { theme } from './theme/theme';
 
 export const App = () => {
-  const routes = useRoutes(Routes);
   return (
     <ThemeProvider theme={theme}>
-      <Header></Header>
-      <Router>{routes}</Router>
+      <Header />
+      <Router />
     </ThemeProvider>
   );
 };
