+++ src/components/Header/components/Logo/Logo.scss
@@ -3,5 +3,5 @@
   justify-content: flex-start;
   width: 3.125rem;
   height: 3.125rem;
-  margin-right: 1.5re;
+  margin-right: 1.5rem;
 }
