+++ src/components/Header/components/SocialsList/SocialsList.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { Box, Button } from '@mui/material';
-import { socials } from '../../constants';
+import { socials } from '../../constants/socials';
 import './SocialsList.scss';
 
 export const SocialsList = () => (
   <Box className="header-socials">
-    {socials.map(({ icon: Icon, url }) => (
+    {socials.map(({ icon: Icon, url }, index) => (
       <Button
         className="header-socials-item"
-        key={url}
+        key={index}
         href={url}
         target="_blank"
       >
