--- src/components/Header/constants/index.ts
@@ -1,2 +1,0 @@
-export { pages } from './pages';
-export { socials } from './socials';
