+++ src/components/Header/constants/socials.ts
@@ -1,5 +1,10 @@
 import { Instagram, LinkedIn, GitHub } from '@mui/icons-material';
-import { SocialsUrl } from '../socialsUrl';
+
+export const enum SocialsUrl {
+  INSTAGRAM = 'http://instagram.com/nasa',
+  LINLEDIN = 'https://www.linkedin.com/company/softserve/',
+  GITHUB = 'https://github.com/kerrimov/the-ssspace',
+}
 
 export const socials = [
   { icon: Instagram, url: SocialsUrl.INSTAGRAM },
