--- src/components/Header/socialsUrl.ts
@@ -1,5 +1,0 @@
-export const enum SocialsUrl {
-  INSTAGRAM = 'http://instagram.com/nasa',
-  LINLEDIN = 'https://www.linkedin.com/company/softserve/',
-  GITHUB = 'https://github.com/kerrimov/the-ssspace',
-}
