+++ src/components/Router/Router.tsx
@@ -1,6 +1,9 @@
 import React from 'react';
 import './Router.scss';
+import { useRoutes } from 'react-router-dom';
+import { Routes } from './Routes';
 
-export const Router: React.FC = ({ children }) => (
-  <div className="router-elements-wrapper">{children}</div>
-);
+export const Router = () => {
+  const routes = useRoutes(Routes);
+  return <div className="router-elements-wrapper">{routes}</div>;
+};
