--- src/components/Routes/index.ts
@@ -1,1 +1,0 @@
-export { Routes } from './Routes';
